#739 [REFACTOR] url 링크를 a 태그로 감싸는 로직 utils로 분리
