+++ src/components/Url/Url.module.css
@@ -1,0 +1,7 @@
+.link {
+  color: var(--blue-3);
+}
+
+.link:hover {
+  text-decoration: underline;
+}
