+++ src/pages/PostPage/PostPage/PostContent.module.css
@@ -3,9 +3,6 @@
   color: var(--gray-4);
   font-size: 0.875rem;
   padding: 0 0.8125rem 0.9375rem 0;
-  white-space: pre-wrap; /* 줄바꿈과 공백을 유지 */
-}
-
-.link {
-  color: var(--blue-3);
+  white-space: pre-wrap;
+  word-break: keep-all;
 }
