+++ src/utils/index.js
@@ -6,6 +6,7 @@ export * from './filterComment.js';
 export * from './findRoute.js';
 export * from './formattedNowTime.js';
 export * from './getBoardTextId.js';
+export * from './link.js';
 export * from './pagination.js';
 export * from './timeAgo.js';
 export * from './validate.js';
